git symbolic-ref: refs/tags/head
Providing this in the git template is equivalent to running:

  git symbolic-ref refs/tags/head HEAD

It allows me to type "head" rather than "HEAD", even on a case-sensitive
file-system. I find this much more convenient for commands like:

  git rebase -i head~4

Note that I originally used refs/heads/head, but by treating head as a
branch rather than a tag, if you check it out (which would be silly to
do, but nonetheless...) git will attach HEAD to it. Attempting to commit
will fail because it's trying to apply two updates to HEAD (one to move
it to the branch tip and another via the ref).