register-sumall-argcomplete
The sumall-cli completions only work if you've installed python's
argcomplete library into your user environment and use it to register
the sumall completions in your shell.

The Nix derivation I wrote to package sumall-cli requires the
argcomplete library, but that doesn't expose argcomplete's executables
in the installed user environment.

I was previously solving this by installing argcomplete into my user
environment separately (as an application, not a library).

Another solution was to list argcomplete under `propagatedUserEnvPkgs`
in the sumall-cli derivation.

But since I'm only interested in using argcomplete for the sumall-cli,
I'd rather not install it's executables directly into my user
environment. Instead, I added a postFix build phase that uses
`makeWrapper` to wrap a call to:

  register-python-argcomplete sumall

in an executable called:

  register-sumall-argcomplete