git config: sign commits with GPG key
I really like the pass (https://www.passwordstore.org) cli pasword
manager, and it makes use of GPG for encryption and signatures.

Following the instructions at,
  https://help.github.com/articles/generating-a-new-gpg-key
I ran `gpg --full-generate-key` and selected the following options:
  key type: RSA and RSA
  key size: 4096 bits
  user id: "Ivan Brennan <ivan.brennan@gmail.com>"
I protected the secret key with a passphrase.

To add the public key to GitHub, I ran:

  gpg --list-secret-keys --with-colons \
    | awk --field-separator=: '/^sec/ { print $5 }' \
    | xargs gpg --export --armor \
    | xsel -ib
  (chromium-browser https://github.com/settings/keys >/dev/null 2>&1 &)

then clicked "New GPG key" and pasted the public key into the text
field.

To transfer the key-pair as well as the GPG config and database to
another machine, simply copy the ~/.gnupg directory.

To transfer just the GPG key-pair to another machine, run the following
on the source machine (adjust DEST to reference the external media --
in this case /run/media/ivan/8383-4531 is a thumb drive I'm using to
perform the transfer):

  DEST=/run/media/ivan/8383-4531/gpg/keys
  KEYID=$(gpg --list-secret-keys --with-colons \
          | awk --field-separator=: '/^sec/ { print $5 }')
  gpg --export --armor $KEYID > $DEST/public.asc
  gpg --export-secret-keys --armor $KEYID > $DEST/secret.asc
  unset DEST KEYID

On the destination machine:

  gpg --import public.asc
  gpg --import secret.asc
  KEYID=$(gpg --list-secret-keys --with-colons \
          | awk --field-separator=: '/^sec/ { print $5 }')
  gpg --edit-key $KEYID
  # command "trust", value "ultimately"
  unset KEYID