shellcheck gsettings script
Run the shellcheck utility and take most (but not all) of its advice.
I'm ignoring its warning about unused variables, as they're actually
variables that are used indirectly via dynamic scoping. Resorting to
dynamically-scoped trickery might be a sign that Bash is the wrong tool
for the job, but this works well enough for now.