bin/dots (using GNU Stow)
Centralize this a bit. I tried housing this in a Nix package, but that's
problematic, since the stow directory will change any time the Nix
package is updated. Then Stow refuses to manage the symlinks that point
to the old version. That also means I could end up with dangling
symlinks if GC cleans out an old store path and a symlink is still
referencing it.